Fix clang warnings with glibc 2.25
When building with `-pedantic` (as neovim does),
clang complains about how `assert()` - among other things - is defined.
The Clang devs say that this is a glibc issue; see the thread at
http://lists.llvm.org/pipermail/cfe-dev/2017-April/053382.html for all
the fun details.

At any rate, these problems go away when building with `-std=c99`
instead of `-std=gnu99`. When doing so, `-D_DEFAULT_SOURCE` is also
needed. See:
http://man7.org/linux/man-pages/man7/feature_test_macros.7.html
https://github.com/neovim/neovim/issues/4042
https://github.com/libuv/libuv/issues/891

As a side note, if we do define `-D_DEFAULT_SOURCE`, is
https://github.com/neovim/neovim/pull/4094 still needed?